+++ config/environment.js
@@ -220,6 +220,7 @@ module.exports = function(environment) {
                 'settings.profile.employment': 'ember_user_settings_profile_page',
                 'settings.profile.name': 'ember_user_settings_profile_page',
                 'settings.profile.social': 'ember_user_settings_profile_page',
+                'settings.account': 'ember_user_settings_account_page',
                 'settings.tokens': 'ember_user_settings_tokens_page',
                 'settings.tokens.index': 'ember_user_settings_tokens_page',
                 'settings.tokens.create': 'ember_user_settings_tokens_page',
@@ -240,6 +241,7 @@ module.exports = function(environment) {
                 institutions: 'institutions_nav_bar',
             },
             storageI18n: 'storage_i18n',
+            enableInactiveSchemas: 'enable_inactive_schemas',
             verifyEmailModals: 'ember_verify_email_modals',
         },
         gReCaptcha: {