+++ ember-cli-build.js
@@ -1,7 +1,6 @@
 /* eslint-env node */
 
 const EmberApp = require('ember-cli/lib/broccoli/ember-app');
-const Funnel = require('broccoli-funnel');
 
 const { EMBER_ENV } = process.env;
 const useCdn = EMBER_ENV === 'production';
@@ -26,10 +25,13 @@ module.exports = function(defaults) {
             includeHighlightJS: false,
             includeFileExtensionInSnippetNames: false,
             snippetSearchPaths: ['lib/handbook/addon'],
-            snippetRegexes: {
+            snippetRegexes: [{
                 begin: /{{#(?:docs-snippet|demo.example|demo.live-example)\sname=(?:"|')(\S+)(?:"|')/,
                 end: /{{\/(?:docs-snippet|demo.example|demo.live-example)}}/,
-            },
+            }, {
+                begin: /<(?:DocsSnippet|demo.example|demo.live-example)\s@name=(?:"|')(\S+)(?:"|')/,
+                end: /<\/(?:DocsSnippet|demo.example|demo.live-example)>/,
+            }],
         };
     }
 
@@ -69,7 +71,7 @@ module.exports = function(defaults) {
         },
         sassOptions: {
             includePaths: [
-                'node_modules/@centerforopenscience/osf-style/sass',
+                'app/styles/',
             ],
         },
         cssModules: {
@@ -152,13 +154,5 @@ module.exports = function(defaults) {
             using: [{ transformation: 'amd', as: 'highlight.js' }],
         });
     }
-
-    const assets = [
-        new Funnel('node_modules/@centerforopenscience/osf-style/img', {
-            srcDir: '/',
-            destDir: 'img',
-        }),
-    ];
-
-    return app.toTree(assets);
+    return app.toTree();
 };