+++ lib/collections/addon/guid/edit/controller.ts
@@ -13,9 +13,16 @@ export default class GuidEdit extends Controller {
     @alias('model.taskInstance.value.collectedMetadatum') collectedMetadatum!: CollectedMetadatum;
     @alias('model.taskInstance.value.collectionItem') collectionItem!: Node;
 
+    isPageDirty: boolean = false;
+
     @action
     returnToProjectOverviewPage() {
         // change to using `transitionToRoute()` once the project overview page is in Ember
         window.location.href = `${config.OSF.url}${this.collectionItem.id}`;
+    }
+
+    @action
+    setPageDirty() {
+        this.set('isPageDirty', true);
     }
 }